+++ src/Symfony/Component/Workflow/Definition.php
@@ -12,6 +12,8 @@
 namespace Symfony\Component\Workflow;
 
 use Symfony\Component\Workflow\Exception\LogicException;
+use Symfony\Component\Workflow\Metadata\InMemoryMetadataStore;
+use Symfony\Component\Workflow\Metadata\MetadataStoreInterface;
 
 /**
  * @author Fabien Potencier <fabien@symfony.com>
@@ -23,13 +25,14 @@ final class Definition
     private $places = array();
     private $transitions = array();
     private $initialPlace;
+    private $metadataStore;
 
     /**
      * @param string[]     $places
      * @param Transition[] $transitions
      * @param string|null  $initialPlace
      */
-    public function __construct(array $places, array $transitions, string $initialPlace = null)
+    public function __construct(array $places, array $transitions, string $initialPlace = null, MetadataStoreInterface $metadataStore = null)
     {
         foreach ($places as $place) {
             $this->addPlace($place);
@@ -40,6 +43,8 @@ final class Definition
         }
 
         $this->setInitialPlace($initialPlace);
+
+        $this->metadataStore = $metadataStore ?: new InMemoryMetadataStore();
     }
 
     /**
@@ -64,6 +69,11 @@ final class Definition
     public function getTransitions(): array
     {
         return $this->transitions;
+    }
+
+    public function getMetadataStore(): MetadataStoreInterface
+    {
+        return $this->metadataStore;
     }
 
     private function setInitialPlace(string $place = null)