+++ src/Symfony/Component/Workflow/CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
  * Deprecated the class `ClassInstanceSupportStrategy` in favor of the class `InstanceOfSupportStrategy`.
  * Added TransitionBlockers as a way to pass around reasons why exactly
    transitions can't be made.
+ * Added a `MetadataStore`.
 
 4.0.0
 -----